[Renderer] Refactoring rendering pipeline [Editor] Add default material when import material is disabled

+++ Source/Framework/Core/Components/TeCCamera.h
@@ -125,10 +125,10 @@ namespace te
         INT32 GetPriority() const { return _internal->GetPriority(); }
 
         /** @copydoc Camera::SetLayers */
-        void SetLayers(UINT64 layers) { _internal->SetLayers(layers); }
+        void SetLayers(UINT32 layers) { _internal->SetLayers(layers); }
 
         /** @copydoc Camera::GetLayers */
-        UINT64 GetLayers() const { return _internal->GetLayers(); }
+        UINT32 GetLayers() const { return _internal->GetLayers(); }
 
         /** @copydoc Camera::SetMSAACount */
         void SetMSAACount(UINT32 count) { _internal->SetMSAACount(count); }
